fix: rollback stats preloading and restore parallel requests
