+++ src/app/api/stats/route.ts
@@ -1,14 +1,29 @@
 import { cookies } from 'next/headers';
 import { NextResponse } from 'next/server';
 import { queryD1 } from '@/utils/d1-util';
-import { getVisitorStats } from '@/utils/stats-util';
 
 const COUNT_INTERVAL_MS = 15 * 60 * 1000;
 const LAST_COUNTED_COOKIE = 'stats_last_counted_at';
 
+interface CountRow {
+  total: number;
+}
+
 export async function GET() {
   try {
-    return NextResponse.json(await getVisitorStats());
+    const today = new Date().toISOString().slice(0, 10);
+
+    const [todayRows, totalRows] = await Promise.all([
+      queryD1<CountRow>('SELECT COALESCE(SUM(count), 0) as total FROM page_views WHERE date = ?', [
+        today,
+      ]),
+      queryD1<CountRow>('SELECT COALESCE(SUM(count), 0) as total FROM page_views', []),
+    ]);
+
+    return NextResponse.json({
+      today: todayRows[0]?.total ?? 0,
+      total: totalRows[0]?.total ?? 0,
+    });
   } catch {
     return NextResponse.json({ today: 0, total: 0 }, { status: 500 });
   }
